+++ app/src/main/java/com/github/sdp/tarjetakuna/database/FBCardPossesion.kt
@@ -3,5 +3,5 @@ package com.github.sdp.tarjetakuna.database
 enum class FBCardPossesion {
     OWNED,
     WANTED,
-    NOTHING
+    NONE
 }
