+++ app/src/main/java/com/github/sdp/tarjetakuna/database/FBMagicCard.kt
@@ -2,6 +2,5 @@ package com.github.sdp.tarjetakuna.database
 
 import com.github.sdp.tarjetakuna.model.MagicCard
 
-data class FBMagicCard(private val card: MagicCard, private val possession: FBCardPossesion) {
-
+data class FBMagicCard(val card: MagicCard, val possession: FBCardPossesion) {
 }
