+++ app/src/main/java/com/github/sdp/tarjetakuna/ui/webapi/magicApi/APIMagicCard.kt
@@ -3,7 +3,7 @@ package com.github.sdp.tarjetakuna.ui.webapi.magicApi
 /**
  * Data class for a MagicCard (need to match the doc in api.magicthegathering.io/v1/)
  */
-data class MagicCard(
+data class APIMagicCard(
     val name: String,
     val manaCost: String,
     val cmc: Int,
