+++ app/src/test/java/com/github/sdp/tarjetakuna/model/FilterTest.kt
@@ -5,10 +5,12 @@ import org.junit.Test
 
 class FilterTest {
 
-    private val card1 = MagicCard("Angel of Mercy", "Flying",
+    private val card1 = MagicCard(
+        "Angel of Mercy", "Flying",
         MagicLayout.Normal, 7, "{5}{W}{W}",
         MagicSet("MT15", "Magic 2015"), 56,
-        "https://gatherer.wizards.com/Handlers/Image.ashx?multiverseid=149935&type=card")
+        "https://gatherer.wizards.com/Handlers/Image.ashx?multiverseid=149935&type=card"
+    )
 
     @Test
     fun filterInitializedCorrectly() {
@@ -20,7 +22,8 @@ class FilterTest {
 
     @Test
     fun filterInitializedCorrectly2() {
-        val magicLayoutArray = arrayListOf(MagicLayout.Normal, MagicLayout.Leveler, MagicLayout.Aftermath)
+        val magicLayoutArray =
+            arrayListOf(MagicLayout.Normal, MagicLayout.Leveler, MagicLayout.Aftermath)
         val cmcArray = arrayListOf(1, 5, 7, 9)
         val filter = Filter("Angel of Mercy", magicLayoutArray, cmcArray)
         assert(filter.name == "Angel of Mercy")
@@ -32,7 +35,8 @@ class FilterTest {
 
     @Test
     fun filterDoesContainWorksCorrectly() {
-        val magicLayoutArray = arrayListOf(MagicLayout.Normal, MagicLayout.Leveler, MagicLayout.Aftermath)
+        val magicLayoutArray =
+            arrayListOf(MagicLayout.Normal, MagicLayout.Leveler, MagicLayout.Aftermath)
         val cmcArray = arrayListOf(1, 5, 7, 9)
         val filter = Filter("Angel of Mercy", magicLayoutArray, cmcArray)
         assert(filter.doesContain(card1))
